Fixed major error with double submission due to progress spinner. Added progress spinner everywhere where load happens. Fixed positioning of progress spinner. Updated shared Food Listing and Delivery objects. Updated retrieval of Food Listing data so users can see more information pertaining to Donors, Receivers, and Deliveries.

+++ server/database/tables/food-listing/food-listing.sql
@@ -28,7 +28,7 @@ ALTER TABLE FoodListing ADD COLUMN IF NOT EXISTS availableUnitsCount            
 ALTER TABLE FoodListing ADD COLUMN IF NOT EXISTS unitsLabel                     TEXT            DEFAULT NULL;
 
 -- Contains the total weight of the entire Food Listing in pounds.
-ALTER TABLE FoodListing ADD COLUMN IF NOT EXISTS totalWeight                    INTEGER         DEFAULT NULL;
+ALTER TABLE FoodListing ADD COLUMN IF NOT EXISTS totalWeight                    NUMERIC(6, 2)   DEFAULT NULL;
 
 -- The recommended vehicle type (size) for food delivery.
 ALTER TABLE FoodListing ADD COLUMN IF NOT EXISTS vehicleType                    VehicleType     NOT NULL DEFAULT 'Sedan';
